Fix bug when reading very large DBF files due to 32bit integer overflow
when reading very large DBF files > 2.4GB due to 32bit integer overflow will attempt to seek to before start of file

+++ EGIS.ShapeFileLib/ShapeFile.cs
@@ -12355,7 +12355,7 @@ namespace EGIS.ShapeFileLib
                 throw new ArgumentException("fieldIndex must be <= DBFRecordHeader.NumFields and >=0");
             }
 			
-			int FieldOffset = this.dBFRecordHeader.HeaderLength + this.dBFRecordHeader.RecordLength*recordNumber + this.dBFRecordHeader.GetFieldDescriptions()[fieldIndex].RecordOffset;
+			long FieldOffset = this.dBFRecordHeader.HeaderLength + this.dBFRecordHeader.RecordLength*(long)recordNumber + this.dBFRecordHeader.GetFieldDescriptions()[fieldIndex].RecordOffset;
 			//string strField;
 			dbfFileStream.Seek(FieldOffset,SeekOrigin.Begin);
 			byte[] data = new byte[this.dBFRecordHeader.GetFieldDescriptions()[fieldIndex].FieldLength];
@@ -12406,7 +12406,7 @@ namespace EGIS.ShapeFileLib
 				throw new ArgumentException("recordNumber must be <= DBFRecordHeader.NumRecords and >=0");
 			}
 			
-			int RecordOffset = this.dBFRecordHeader.HeaderLength + this.dBFRecordHeader.RecordLength*recordNumber;
+			long RecordOffset = this.dBFRecordHeader.HeaderLength + this.dBFRecordHeader.RecordLength* (long)recordNumber;
 			
 			int numFields = this.dBFRecordHeader.FieldCount;
 			string[] strFields = new string[numFields];
@@ -12462,7 +12462,7 @@ namespace EGIS.ShapeFileLib
             List<string> records = new List<string>();
             Dictionary<string, int> d = new Dictionary<string, int>();
 
-            int FieldOffset = this.dBFRecordHeader.HeaderLength + this.dBFRecordHeader.GetFieldDescriptions()[fieldIndex].RecordOffset;
+            long FieldOffset = this.dBFRecordHeader.HeaderLength + this.dBFRecordHeader.GetFieldDescriptions()[fieldIndex].RecordOffset;
             
             int fieldLength = this.dBFRecordHeader.GetFieldDescriptions()[fieldIndex].FieldLength;
             dbfFileStream.Seek(FieldOffset, SeekOrigin.Begin);
